First drop of kerberos support

+++ pkg/controllers/obj_reconcile_test.go
@@ -23,11 +23,12 @@ import (
 	vapi "github.com/vertica/vertica-kubernetes/api/v1beta1"
 	"github.com/vertica/vertica-kubernetes/pkg/cmds"
 	"github.com/vertica/vertica-kubernetes/pkg/names"
+	"github.com/vertica/vertica-kubernetes/pkg/paths"
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/kubernetes/scheme"
 	ctrl "sigs.k8s.io/controller-runtime"
 	//+kubebuilder:scaffold:imports
 )
@@ -38,18 +39,23 @@ import (
 var _ = Describe("obj_reconcile", func() {
 	ctx := context.Background()
 
-	createCrd := func(vdb *vapi.VerticaDB) {
+	runReconciler := func(vdb *vapi.VerticaDB, expResult ctrl.Result) {
+		// Create any dependent objects for the CRD.
+		pfacts := MakePodFacts(k8sClient, &cmds.FakePodRunner{})
+		objr := MakeObjReconciler(vrec, logger, vdb, &pfacts)
+		Expect(objr.Reconcile(ctx, &ctrl.Request{})).Should(Equal(expResult))
+	}
+
+	createCrd := func(vdb *vapi.VerticaDB, doReconcile bool) {
 		Expect(k8sClient.Create(ctx, vdb)).Should(Succeed())
 		nameLookup := vdb.ExtractNamespacedName()
 
 		createdVdb := &vapi.VerticaDB{}
 		Expect(k8sClient.Get(ctx, nameLookup, createdVdb)).Should(Succeed())
 
-		// Create any dependent objects for the CRD.
-		pfacts := MakePodFacts(k8sClient, &cmds.FakePodRunner{})
-		objr := MakeObjReconciler(k8sClient, scheme.Scheme, logger, createdVdb, &pfacts)
-		_, err := objr.Reconcile(ctx, &ctrl.Request{})
-		Expect(err).Should(Succeed())
+		if doReconcile {
+			runReconciler(vdb, ctrl.Result{})
+		}
 	}
 
 	deleteCrd := func(vdb *vapi.VerticaDB) {
@@ -68,26 +74,38 @@ var _ = Describe("obj_reconcile", func() {
 
 		for i := range vdb.Spec.Subclusters {
 			svc := &corev1.Service{}
-			Expect(k8sClient.Get(ctx, names.GenExtSvcName(vdb, &vdb.Spec.Subclusters[i]), svc)).Should(Succeed())
-			Expect(svc.ObjectMeta.OwnerReferences).To(ContainElement(expOwnerRef))
-			Expect(k8sClient.Delete(ctx, svc)).Should(Succeed())
+			err := k8sClient.Get(ctx, names.GenExtSvcName(vdb, &vdb.Spec.Subclusters[i]), svc)
+			if err == nil {
+				Expect(svc.ObjectMeta.OwnerReferences).To(ContainElement(expOwnerRef))
+				Expect(k8sClient.Delete(ctx, svc)).Should(Succeed())
+			} else {
+				Expect(errors.IsNotFound(err)).Should(BeTrue())
+			}
 
 			stsNm := names.GenStsName(vdb, &vdb.Spec.Subclusters[i])
 			sts := &appsv1.StatefulSet{}
-			Expect(k8sClient.Get(ctx, stsNm, sts)).Should(Succeed())
-			Expect(sts.ObjectMeta.OwnerReferences).To(ContainElement(expOwnerRef))
-			Expect(k8sClient.Delete(ctx, sts)).Should(Succeed())
+			err = k8sClient.Get(ctx, stsNm, sts)
+			if err == nil {
+				Expect(sts.ObjectMeta.OwnerReferences).To(ContainElement(expOwnerRef))
+				Expect(k8sClient.Delete(ctx, sts)).Should(Succeed())
+			} else {
+				Expect(errors.IsNotFound(err)).Should(BeTrue())
+			}
 		}
 		svc := &corev1.Service{}
-		Expect(k8sClient.Get(ctx, names.GenHlSvcName(vdb), svc)).Should(Succeed())
-		Expect(svc.ObjectMeta.OwnerReferences).To(ContainElement(expOwnerRef))
-		Expect(k8sClient.Delete(ctx, svc)).Should(Succeed())
+		err := k8sClient.Get(ctx, names.GenHlSvcName(vdb), svc)
+		if err == nil {
+			Expect(svc.ObjectMeta.OwnerReferences).To(ContainElement(expOwnerRef))
+			Expect(k8sClient.Delete(ctx, svc)).Should(Succeed())
+		} else {
+			Expect(errors.IsNotFound(err)).Should(BeTrue())
+		}
 	}
 
 	Context("When reconciling a VerticaDB CRD", func() {
 		It("Should create service objects", func() {
 			vdb := vapi.MakeVDB()
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			By("Checking the VerticaDB has an external service object")
@@ -117,7 +135,7 @@ var _ = Describe("obj_reconcile", func() {
 			vdb.Spec.Subclusters[0].NodePort = desiredNodePort
 			vdb.Spec.Subclusters[0].ExternalIPs = desiredExternalIPs
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			extNameLookup := names.GenExtSvcName(vdb, &vdb.Spec.Subclusters[0])
@@ -138,7 +156,7 @@ var _ = Describe("obj_reconcile", func() {
 
 			// Refresh any dependent objects
 			pfacts := MakePodFacts(k8sClient, &cmds.FakePodRunner{})
-			objr := MakeObjReconciler(k8sClient, scheme.Scheme, logger, vdb, &pfacts)
+			objr := MakeObjReconciler(vrec, logger, vdb, &pfacts)
 			_, err := objr.Reconcile(ctx, &ctrl.Request{})
 			Expect(err).Should(Succeed())
 
@@ -164,7 +182,7 @@ var _ = Describe("obj_reconcile", func() {
 				}
 			}
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			svc := &corev1.Service{}
@@ -183,7 +201,7 @@ var _ = Describe("obj_reconcile", func() {
 			var desiredSize int32 = 16
 			vdb.Spec.Subclusters[0].Size = desiredSize
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -198,7 +216,7 @@ var _ = Describe("obj_reconcile", func() {
 			vdb := vapi.MakeVDB()
 			vdb.Spec.ImagePullPolicy = corev1.PullNever
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -213,7 +231,7 @@ var _ = Describe("obj_reconcile", func() {
 			desiredStorageClass := "my-storage"
 			vdb.Spec.Local.StorageClass = desiredStorageClass
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -231,7 +249,7 @@ var _ = Describe("obj_reconcile", func() {
 			}
 			vdb.Spec.Subclusters[0].NodeSelector = desiredNodeSelector
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -270,7 +288,7 @@ var _ = Describe("obj_reconcile", func() {
 			}
 			vdb.Spec.Subclusters[0].Affinity = desiredAffinity
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -291,7 +309,7 @@ var _ = Describe("obj_reconcile", func() {
 			}
 			vdb.Spec.Subclusters[0].Tolerations = desiredTolerations
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -311,7 +329,7 @@ var _ = Describe("obj_reconcile", func() {
 			}
 			vdb.Spec.Subclusters[0].Resources = desiredResources
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -328,7 +346,7 @@ var _ = Describe("obj_reconcile", func() {
 				Size: 8,
 			})
 
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -341,7 +359,7 @@ var _ = Describe("obj_reconcile", func() {
 
 		It("Increasing the size of the subcluster should cause the sts to scale out", func() {
 			vdb := vapi.MakeVDB()
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 			origSize := vdb.Spec.Subclusters[0].Size
 
@@ -357,7 +375,7 @@ var _ = Describe("obj_reconcile", func() {
 
 			// Refresh any dependent objects
 			pfacts := MakePodFacts(k8sClient, &cmds.FakePodRunner{})
-			objr := MakeObjReconciler(k8sClient, scheme.Scheme, logger, vdb, &pfacts)
+			objr := MakeObjReconciler(vrec, logger, vdb, &pfacts)
 			_, err := objr.Reconcile(ctx, &ctrl.Request{})
 			Expect(err).Should(Succeed())
 
@@ -369,7 +387,7 @@ var _ = Describe("obj_reconcile", func() {
 		It("should have updateStrategy OnDelete for kSafety 0", func() {
 			vdb := vapi.MakeVDB()
 			vdb.Spec.KSafety = vapi.KSafety0
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			updateStrategyHelper(ctx, vdb, appsv1.OnDeleteStatefulSetStrategyType)
@@ -378,7 +396,7 @@ var _ = Describe("obj_reconcile", func() {
 		It("should have updateStrategy RollingUpdate for kSafety 1", func() {
 			vdb := vapi.MakeVDB()
 			vdb.Spec.KSafety = vapi.KSafety1
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			updateStrategyHelper(ctx, vdb, appsv1.RollingUpdateStatefulSetStrategyType)
@@ -399,7 +417,7 @@ var _ = Describe("obj_reconcile", func() {
 					Requests: corev1.ResourceList{"cpu": cpuResource, "memory": memResource},
 				},
 			})
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -416,7 +434,7 @@ var _ = Describe("obj_reconcile", func() {
 			vdb := vapi.MakeVDB()
 			const PullSecretName = "docker-info"
 			vdb.Spec.ImagePullSecrets = append(vdb.Spec.ImagePullSecrets, corev1.LocalObjectReference{Name: PullSecretName})
-			createCrd(vdb)
+			createCrd(vdb, true)
 			defer deleteCrd(vdb)
 
 			sts := &appsv1.StatefulSet{}
@@ -424,6 +442,55 @@ var _ = Describe("obj_reconcile", func() {
 			Expect(k8sClient.Get(ctx, nm, sts)).Should(Succeed())
 			Expect(len(sts.Spec.Template.Spec.ImagePullSecrets)).Should(Equal(1))
 			Expect(sts.Spec.Template.Spec.ImagePullSecrets).Should(ContainElement(vdb.Spec.ImagePullSecrets[0]))
+		})
+
+		It("should requeue if the license secret is not found", func() {
+			vdb := vapi.MakeVDB()
+			vdb.Spec.LicenseSecret = "not-here"
+			createCrd(vdb, false)
+			defer deleteCrd(vdb)
+
+			pfacts := MakePodFacts(k8sClient, &cmds.FakePodRunner{})
+			objr := MakeObjReconciler(vrec, logger, vdb, &pfacts)
+			Expect(objr.Reconcile(ctx, &ctrl.Request{})).Should(Equal(ctrl.Result{Requeue: true}))
+		})
+
+		It("should requeue if the kerberos secret is not found", func() {
+			vdb := vapi.MakeVDB()
+			vdb.Spec.KerberosSecret = "not-here"
+			createCrd(vdb, false)
+			defer deleteCrd(vdb)
+
+			pfacts := MakePodFacts(k8sClient, &cmds.FakePodRunner{})
+			objr := MakeObjReconciler(vrec, logger, vdb, &pfacts)
+			Expect(objr.Reconcile(ctx, &ctrl.Request{})).Should(Equal(ctrl.Result{Requeue: true}))
+		})
+
+		It("should succeed if the kerberos secret is setup correctly", func() {
+			vdb := vapi.MakeVDB()
+			vdb.Spec.KerberosSecret = "my-secret-v1"
+			secret := buildKerberosSecretBase(vdb)
+			secret.Data[paths.Krb5Keytab] = []byte("keytab")
+			secret.Data[paths.Krb5Conf] = []byte("conf")
+			Expect(k8sClient.Create(ctx, secret)).Should(Succeed())
+			defer deleteSecret(ctx, vdb, vdb.Spec.KerberosSecret)
+			createCrd(vdb, false)
+			defer deleteCrd(vdb)
+
+			runReconciler(vdb, ctrl.Result{})
+		})
+
+		It("should requeue if the kerberos secret has a missing keytab", func() {
+			vdb := vapi.MakeVDB()
+			vdb.Spec.KerberosSecret = "my-secret-v2"
+			secret := buildKerberosSecretBase(vdb)
+			secret.Data[paths.Krb5Conf] = []byte("conf") // Only the krb5.conf
+			Expect(k8sClient.Create(ctx, secret)).Should(Succeed())
+			defer deleteSecret(ctx, vdb, vdb.Spec.KerberosSecret)
+			createCrd(vdb, false)
+			defer deleteCrd(vdb)
+
+			runReconciler(vdb, ctrl.Result{Requeue: true})
 		})
 	})
 })
